Use jasmine testing framework for unit tests
Jasmine framework has been chosen as it is possible to easy setup
browser tests if it is necessary.

Other frameworks should also work fine, but there doesn't seem to be
much difference at the moment. We also can switch to another framework
in future.
