+++ javascript/spec/data/curtain-1.bin
